requirements.txt
Flask==3.1.1
Jinja2==3.1.2
numpy==2.3.1
pandas==2.3.1
peewee==3.17.0
plotly==5.18.0
PyYAML==6.0.2
reportlab==4.2.0
weasyprint==65.1
Werkzeug==3.1.3
gunicorn==22.0.0
Flask-Session
psycopg2-binary
python-dotenv
